requirements.txt
bcrypt==4.1.2
django==4.2.8
djangorestframework==3.14.0
djangorestframework-api-key==3.0.0
django-health-check==3.17.0
elastic-apm==6.19.0
gunicorn==21.2.0
psycopg==3.1.14
python-decouple==3.8
PyJWT[crypto]==2.8.0
